Reimplement RunLast event into more generic Anchor event
The Anchor event class generally anchors a single coro function and
other tasks added to the event are run in relation to the anchor. This
allows not only re-implementing the RunLast event as a sub-class of
Anchor, but also allows implementing RunFirst ie the anchor coro
function is run and completed first before the other tasks added to the
event are run.