FIX: Rename LoadLimitEvent attr _kwargs to _result
This is simply a cleanup to ensure names are clearer. The original
intent behind LoadLimitEvent._kwargs is that it hold a future for the
event which holds values passed in from the call of set() ie set() takes
kwargs, and this kwargs is the result of the future.

This rationale, however, isn't apparent and thus the rename to something
clearer.