New coro funcs: flushtosql(), flushtosql_shutdown()
flushtosql() takes all data from a Period stats dict and stores it in an
sql db. flushtosql_shutdown() is meant to be run just before loop
shutdown, flushing any remaining data in the stats dict to the sql db.

flushtosql() is meant to be attached to the recordperiod event.

* New dependency: SQLAlchemy
* New coro func: flushtosql()
* New coro func: flushtosql_shutdown()

Note: it is recommended that flushtosql() be run as a non-scheduled coro
func. If run as a scheduled task, it is possible it will not receive all
data generated from the coro funcs setting the recordperiod event.